chore: organize root directory structure for GitHub release
- Move Python test files to tests/python/
- Move Rust test files to tests/rust/ (keep only build.rs in root)
- Move benchmark and analysis tools to misc/tools/
- Move logs to misc/logs/
- Move JSON result files to reports/benchmarks/ and reports/test-results/
- Move data files to misc/data/
- Clean up executables and temporary files
- Optimize root directory for professional GitHub release
