--- Test.atxt
@@ -1,15 +1,0 @@
-,-----------.  ,-----------.
-|gRPC Server|  |REST Server|
-|-----------|  |-----------|
-`-----------'  `-----------'
-                            
-                            
-    ,------------------.    
-    |Collection Service|    
-    |------------------|    
-    `------------------'    
-              |             
-       ,------------.       
-       |Avro Service|       
-       |------------|       
-       `------------'
