--- addonindexing.txt
@@ -1,108 +1,0 @@
- Let's Design: Hybrid Index Structure for Adaptive Vector DB
-➕ Introducing: Unified Sparse-Dense Hybrid Index (USHDI)
-This is a layered indexing design that integrates:
-
-Sparse vector lookup
-
-Dense vector grouping
-
-Metadata filters
-
-ID-based joins
-
-ANN hooks
-
-🧱 Architecture of USHDI
-✅ Core Components:
-1. Global ID Index (Trie + Hash Map)
-Map: id → {location: (partition_id, offset_in_file)}
-
-Trie prefix support allows prefix queries
-
-Supports join between metadata and vector store
-
-2. Metadata Index (Segmented Column + Bitmap Filters)
-Columnar metadata stored in Parquet
-
-Augmented with Roaring Bitmaps or dictionary-encoded indexes for fast filtering
-
-Filters like metadata.language = "en" map to bitmap → row IDs
-
-3. Dense Vector Index (Parquet Row Group + HNSW / PQ)
-Indexes built per partition
-
-Store ANN index pointer alongside row group offset
-
-For ANN queries, we use partitioning model → ANN index → row ID
-
-4. Sparse Vector Index (Compressed Sparse Row + LSM + Sketch Tables)
-LSM-tree stores ID → sparse vector
-
-Optional Count-Min Sketch or SimHash to support approximate similarity filtering
-
-Could support ANN over sparse vectors via dedicated structures like MinHash LSH
-
-5. Join Engine (RowSet Intersection + Bloom Cache)
-Result sets from metadata filters, ANN search, or ID lookups are joined using:
-
-Row ID intersection
-
-Bloom filters for false-positive fast rejection
-
-Priority queue for relevance re-ranking
-
-🔄 Insert Workflow (Dense & Sparse)
-text
-Copy
-Edit
-→ WAL
-→ Determine vector type (dense/sparse)
-→ Write:
-   - Metadata to Parquet
-   - Dense vector to row group + ANN build queue
-   - Sparse vector to LSM or KV
-→ Update:
-   - ID index
-   - Metadata bitmap
-   - ANN index queue
-🔍 Query Workflow Example
-Query:
-"Give me vectors where metadata.user_type = 'pro' and similarity > 0.9 to this vector"
-
-Steps:
-Filter metadata bitmap → Row IDs
-
-ANN index query in relevant partitions → Row IDs
-
-Intersect row IDs
-
-Fetch vector from Parquet or KV using global ID map
-
-Return ranked results
-
-🧪 Data Structure Design Summary
-Component	Backing DS	Purpose
-ID Index	Trie + Hash Map	Fast global lookup
-Metadata Index	Column Store + Bitmap	Efficient predicate filtering
-Dense Vector Index	HNSW / IVF / PQ	ANN retrieval for dense rows
-Sparse Vector Index	LSM + Sparse Format + MinHash	ANN or fast lookup in sparse space
-Join Layer	Bloom Filters + RowSet	Result combination engine
-
-🚀 Optimizations & Extensions
-Multi-level caching (hot vectors in memory)
-
-Streaming mode: index vector insert batches as mini-segments
-
-Reorg tool: rebalance partitions + rebuild ANN periodically
-
-Optional: Time-travel queries using versioned vector IDs or timestamp columns
-
-🧩 Real-World Comparison
-Feature	Pinecone	Qdrant	Milvus	This Design
-Sparse Support	❌	Partial	❌	✅
-Hybrid Storage	❌	❌	❌	✅
-ML-based Partitioning	❌	❌	✅	✅ (dynamic)
-Row-level Metadata Filtering	✅	✅	✅	✅
-Indexed Sparse Vectors	❌	❌	❌	✅ (MinHash / Sketch)
-
-
