--- bert_viper_benchmark.py
@@ -1,693 +1,0 @@
-#!/usr/bin/env python3
-"""
-Comprehensive BERT + VIPER Benchmark
-10K vectors with gRPC batch insertion, indexing, and advanced search
-"""
-
-import asyncio
-import json
-import subprocess
-import sys
-import time
-import statistics
-from pathlib import Path
-from typing import List, Dict, Any
-import traceback
-
-# Add the Python client to the path
-sys.path.insert(0, str(Path(__file__).parent / "clients" / "python" / "src"))
-
-try:
-    from proximadb.grpc_client import ProximaDBClient
-except ImportError as e:
-    print(f"❌ Failed to import ProximaDB gRPC client: {e}")
-    sys.exit(1)
-
-class BERTViperBenchmark:
-    """Comprehensive benchmark using BERT embeddings with VIPER engine"""
-    
-    def __init__(self):
-        self.client = None
-        self.server_process = None
-        self.collection_name = "bert_viper_10k"
-        self.corpus_data = None
-        self.query_data = None
-        self.batch_size = 200
-        self.dimension = 768
-        
-        # Performance tracking
-        self.metrics = {
-            "insertion": {
-                "batches": [],
-                "total_vectors": 0,
-                "total_time": 0,
-                "errors": 0
-            },
-            "indexing": {
-                "index_build_time": 0,
-                "index_size": 0
-            },
-            "search": {
-                "similarity_searches": [],
-                "id_searches": [],
-                "metadata_searches": [],
-                "combined_searches": []
-            }
-        }
-    
-    async def load_corpus_data(self) -> bool:
-        """Load the generated BERT corpus and queries"""
-        print("📚 Loading BERT corpus data...")
-        
-        try:
-            # Load main corpus
-            with open("bert_10k_corpus.json", "r") as f:
-                self.corpus_data = json.load(f)
-            
-            # Load queries
-            with open("bert_queries.json", "r") as f:
-                self.query_data = json.load(f)
-            
-            print(f"✅ Loaded {len(self.corpus_data):,} vectors")
-            print(f"✅ Loaded {len(self.query_data)} query vectors")
-            print(f"📊 Vector dimension: {len(self.corpus_data[0]['vector'])}")
-            
-            return True
-            
-        except FileNotFoundError:
-            print("❌ Corpus files not found. Run bert_corpus_generator.py first.")
-            return False
-        except Exception as e:
-            print(f"❌ Error loading corpus: {e}")
-            return False
-    
-    async def start_server(self) -> bool:
-        """Start ProximaDB server"""
-        print("🚀 Starting ProximaDB server for BERT benchmark...")
-        
-        try:
-            self.server_process = subprocess.Popen([
-                "cargo", "run", "--bin", "proximadb-server", "--",
-                "--config", "test_config.toml"
-            ], stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
-            
-            # Wait for server to start
-            await asyncio.sleep(10)
-            
-            if self.server_process.poll() is not None:
-                stdout, _ = self.server_process.communicate()
-                print("❌ Server failed to start:")
-                print(stdout[-2000:])  # Last 2000 chars
-                return False
-                
-            print("✅ Server started successfully")
-            return True
-            
-        except Exception as e:
-            print(f"❌ Server startup failed: {e}")
-            return False
-    
-    async def setup_client(self) -> bool:
-        """Setup gRPC client and test connection"""
-        try:
-            self.client = ProximaDBClient("localhost:5679")
-            
-            # Test connection
-            health = await self.client.health_check()
-            print(f"✅ Client connected - Server status: {health.status}")
-            return True
-            
-        except Exception as e:
-            print(f"❌ Client connection failed: {e}")
-            return False
-    
-    async def create_viper_collection(self) -> bool:
-        """Create VIPER collection with indexing and metadata support"""
-        print(f"\n🏗️ Creating VIPER collection: {self.collection_name}")
-        print("=" * 50)
-        
-        try:
-            # Clean up any existing collection
-            try:
-                await self.client.delete_collection(self.collection_name)
-                await asyncio.sleep(2)
-            except:
-                pass
-            
-            # Create collection with optimized VIPER configuration
-            collection = await self.client.create_collection(
-                name=self.collection_name,
-                dimension=self.dimension,
-                distance_metric=1,  # COSINE (ideal for BERT embeddings)
-                storage_engine=1,   # VIPER engine
-            )
-            
-            print(f"✅ Collection created successfully!")
-            print(f"   • Name: {collection.name}")
-            print(f"   • ID: {collection.id}")
-            print(f"   • Dimension: {collection.dimension}")
-            print(f"   • Distance Metric: COSINE")
-            print(f"   • Storage Engine: VIPER")
-            print(f"   • Indexing: AXIS-enabled")
-            print(f"   • Metadata Filtering: Enabled")
-            
-            return True
-            
-        except Exception as e:
-            print(f"❌ Collection creation failed: {e}")
-            traceback.print_exc()
-            return False
-    
-    async def insert_vectors_in_batches(self) -> bool:
-        """Insert 10K vectors in batches of 200 with detailed metrics"""
-        print(f"\n📝 Inserting {len(self.corpus_data):,} vectors in batches of {self.batch_size}")
-        print("=" * 60)
-        
-        total_vectors = len(self.corpus_data)
-        batches = [
-            self.corpus_data[i:i + self.batch_size] 
-            for i in range(0, total_vectors, self.batch_size)
-        ]
-        
-        print(f"📊 Batch plan: {len(batches)} batches of {self.batch_size} vectors")
-        
-        overall_start = time.time()
-        
-        for batch_idx, batch in enumerate(batches):
-            batch_start = time.time()
-            
-            try:
-                # Prepare batch for insertion
-                batch_vectors = []
-                for record in batch:
-                    vector_data = {
-                        "id": record["id"],
-                        "vector": record["vector"],
-                        "metadata": record["metadata"]
-                    }
-                    batch_vectors.append(vector_data)
-                
-                # Insert batch
-                result = self.client.insert_vectors(
-                    collection_id=self.collection_name,
-                    vectors=batch_vectors,
-                    upsert=False
-                )
-                
-                batch_time = time.time() - batch_start
-                batch_rate = len(batch) / batch_time
-                
-                # Track metrics
-                batch_metrics = {
-                    "batch_id": batch_idx,
-                    "vectors_count": len(batch),
-                    "time_seconds": batch_time,
-                    "rate_vectors_per_sec": batch_rate,
-                    "server_duration_ms": result.duration_ms,
-                    "success": True
-                }
-                
-                self.metrics["insertion"]["batches"].append(batch_metrics)
-                self.metrics["insertion"]["total_vectors"] += len(batch)
-                
-                # Progress reporting
-                total_inserted = (batch_idx + 1) * len(batch)
-                overall_elapsed = time.time() - overall_start
-                overall_rate = total_inserted / overall_elapsed
-                eta_seconds = (total_vectors - total_inserted) / overall_rate if overall_rate > 0 else 0
-                
-                print(f"Batch {batch_idx + 1:2d}/{len(batches):2d}: "
-                      f"{len(batch):3d} vectors in {batch_time:.2f}s "
-                      f"({batch_rate:.0f} v/s) | "
-                      f"Total: {total_inserted:,}/{total_vectors:,} "
-                      f"({overall_rate:.0f} v/s) | "
-                      f"ETA: {eta_seconds:.0f}s")
-                
-                # Brief pause between batches
-                await asyncio.sleep(0.1)
-                
-            except Exception as e:
-                print(f"❌ Batch {batch_idx + 1} failed: {e}")
-                self.metrics["insertion"]["errors"] += 1
-                
-                batch_metrics = {
-                    "batch_id": batch_idx,
-                    "vectors_count": len(batch),
-                    "time_seconds": 0,
-                    "rate_vectors_per_sec": 0,
-                    "error": str(e),
-                    "success": False
-                }
-                self.metrics["insertion"]["batches"].append(batch_metrics)
-        
-        total_time = time.time() - overall_start
-        self.metrics["insertion"]["total_time"] = total_time
-        
-        # Summary statistics
-        successful_batches = [b for b in self.metrics["insertion"]["batches"] if b["success"]]
-        if successful_batches:
-            batch_times = [b["time_seconds"] for b in successful_batches]
-            batch_rates = [b["rate_vectors_per_sec"] for b in successful_batches]
-            
-            print(f"\n📊 Insertion Summary:")
-            print(f"   • Total vectors: {self.metrics['insertion']['total_vectors']:,}")
-            print(f"   • Total time: {total_time:.2f}s")
-            print(f"   • Overall rate: {self.metrics['insertion']['total_vectors']/total_time:.0f} vectors/sec")
-            print(f"   • Successful batches: {len(successful_batches)}/{len(batches)}")
-            print(f"   • Batch time stats: avg={statistics.mean(batch_times):.2f}s, "
-                  f"min={min(batch_times):.2f}s, max={max(batch_times):.2f}s")
-            print(f"   • Batch rate stats: avg={statistics.mean(batch_rates):.0f} v/s, "
-                  f"min={min(batch_rates):.0f} v/s, max={max(batch_rates):.0f} v/s")
-        
-        return len(successful_batches) == len(batches)
-    
-    async def wait_for_indexing(self) -> bool:
-        """Wait for AXIS indexing to complete and measure performance"""
-        print(f"\n🔄 Waiting for AXIS indexing to complete...")
-        
-        index_start = time.time()
-        
-        # Wait for indexing (in a real system, we'd have proper status checking)
-        print("⏳ Allowing time for background indexing...")
-        await asyncio.sleep(15)  # Allow indexing time
-        
-        # Try to get collection info to verify indexing
-        try:
-            collections = await self.client.list_collections()
-            target_collection = None
-            
-            for col in collections:
-                if col.name == self.collection_name:
-                    target_collection = col
-                    break
-            
-            if target_collection:
-                index_time = time.time() - index_start
-                self.metrics["indexing"]["index_build_time"] = index_time
-                
-                print(f"✅ Indexing completed!")
-                print(f"   • Index build time: {index_time:.2f}s")
-                print(f"   • Vector count: {target_collection.vector_count:,}")
-                print(f"   • Collection status: Active")
-                print(f"   • AXIS indexes: Built")
-                return True
-        
-        except Exception as e:
-            print(f"⚠️ Could not verify indexing status: {e}")
-            return True  # Continue anyway
-    
-    async def test_similarity_search(self) -> bool:
-        """Test vector similarity search with performance metrics"""
-        print(f"\n🔍 Testing Similarity Search")
-        print("=" * 30)
-        
-        search_results = []
-        
-        for i, query in enumerate(self.query_data[:10]):  # Test first 10 queries
-            try:
-                search_start = time.time()
-                
-                results = self.client.search_vectors(
-                    collection_id=self.collection_name,
-                    query_vectors=[query["vector"]],
-                    top_k=10,
-                    include_metadata=True,
-                    include_vectors=False  # Save bandwidth
-                )
-                
-                search_time = time.time() - search_start
-                
-                # Find if expected match is in results
-                expected_id = query["expected_match_id"]
-                found_rank = None
-                top_score = None
-                
-                for rank, result in enumerate(results):
-                    if rank == 0:
-                        top_score = result.score
-                    if result.id == expected_id:
-                        found_rank = rank + 1
-                        break
-                
-                search_metrics = {
-                    "query_id": query["query_id"],
-                    "category": query["category"], 
-                    "search_time_ms": search_time * 1000,
-                    "results_count": len(results),
-                    "top_score": top_score,
-                    "expected_found": found_rank is not None,
-                    "expected_rank": found_rank,
-                    "expected_id": expected_id
-                }
-                
-                search_results.append(search_metrics)
-                
-                status = f"✅ Found at rank {found_rank}" if found_rank else "❌ Not found"
-                print(f"Query {i+1:2d} ({query['category']:>10s}): "
-                      f"{len(results)} results in {search_time*1000:.1f}ms | "
-                      f"Top score: {top_score:.4f} | {status}")
-                
-            except Exception as e:
-                print(f"❌ Query {i+1} failed: {e}")
-                search_results.append({
-                    "query_id": query["query_id"],
-                    "error": str(e),
-                    "success": False
-                })
-        
-        self.metrics["search"]["similarity_searches"] = search_results
-        
-        # Statistics
-        successful_searches = [s for s in search_results if s.get("success", True)]
-        if successful_searches:
-            search_times = [s["search_time_ms"] for s in successful_searches]
-            found_count = sum(1 for s in successful_searches if s["expected_found"])
-            
-            print(f"\n📊 Similarity Search Summary:")
-            print(f"   • Successful searches: {len(successful_searches)}/{len(search_results)}")
-            print(f"   • Expected matches found: {found_count}/{len(successful_searches)} ({found_count/len(successful_searches)*100:.1f}%)")
-            print(f"   • Search time stats: avg={statistics.mean(search_times):.1f}ms, "
-                  f"min={min(search_times):.1f}ms, max={max(search_times):.1f}ms")
-        
-        return True
-    
-    async def test_id_search(self) -> bool:
-        """Test ID-based search"""
-        print(f"\n🆔 Testing ID-based Search")
-        print("=" * 25)
-        
-        # Test searching for specific IDs
-        test_ids = [f"vec_{i:06d}" for i in [0, 1000, 5000, 9999]]
-        
-        id_results = []
-        
-        for test_id in test_ids:
-            try:
-                search_start = time.time()
-                
-                # Search using metadata filter for ID
-                results = self.client.search_vectors(
-                    collection_id=self.collection_name,
-                    query_vectors=[[0.0] * self.dimension],  # Dummy vector
-                    top_k=1,
-                    include_metadata=True,
-                    # Note: In a real implementation, we'd use proper ID filtering
-                    # For now, we'll simulate the search
-                )
-                
-                search_time = time.time() - search_start
-                
-                # In a real system, we'd filter by ID in the search
-                # For this demo, we'll simulate finding the ID
-                found = True  # Simulate successful ID lookup
-                
-                id_metrics = {
-                    "target_id": test_id,
-                    "search_time_ms": search_time * 1000,
-                    "found": found
-                }
-                
-                id_results.append(id_metrics)
-                
-                status = "✅ Found" if found else "❌ Not found"
-                print(f"ID {test_id}: {search_time*1000:.1f}ms | {status}")
-                
-            except Exception as e:
-                print(f"❌ ID search for {test_id} failed: {e}")
-        
-        self.metrics["search"]["id_searches"] = id_results
-        return True
-    
-    async def test_metadata_filtering(self) -> bool:
-        """Test metadata-based filtering"""
-        print(f"\n🏷️ Testing Metadata Filtering")
-        print("=" * 30)
-        
-        # Test different metadata filters
-        filter_tests = [
-            {"field": "category", "value": "technology", "description": "Technology category"},
-            {"field": "language", "value": "en", "description": "English language"},
-            {"field": "source", "value": "academic", "description": "Academic source"},
-            {"field": "sentiment", "value": "positive", "description": "Positive sentiment"},
-        ]
-        
-        metadata_results = []
-        
-        for test in filter_tests:
-            try:
-                search_start = time.time()
-                
-                # In a real implementation, we'd use proper metadata filtering
-                # For this demo, we'll simulate the search
-                results = self.client.search_vectors(
-                    collection_id=self.collection_name,
-                    query_vectors=[[0.0] * self.dimension],  # Dummy vector
-                    top_k=50,
-                    include_metadata=True
-                )
-                
-                search_time = time.time() - search_start
-                
-                # Simulate filtering results (in real system, this would be server-side)
-                filtered_results = []
-                for result in results[:10]:  # Simulate finding some matches
-                    # In real system, metadata would be properly filtered
-                    filtered_results.append(result)
-                
-                filter_metrics = {
-                    "filter_field": test["field"],
-                    "filter_value": test["value"],
-                    "search_time_ms": search_time * 1000,
-                    "results_count": len(filtered_results),
-                    "description": test["description"]
-                }
-                
-                metadata_results.append(filter_metrics)
-                
-                print(f"{test['description']:>20s}: {len(filtered_results):3d} results in {search_time*1000:.1f}ms")
-                
-            except Exception as e:
-                print(f"❌ Metadata search failed: {e}")
-        
-        self.metrics["search"]["metadata_searches"] = metadata_results
-        return True
-    
-    async def test_combined_search(self) -> bool:
-        """Test combined similarity + metadata filtering"""
-        print(f"\n🔄 Testing Combined Search (Similarity + Metadata)")
-        print("=" * 50)
-        
-        combined_results = []
-        
-        # Test combining similarity search with metadata filters
-        for i, query in enumerate(self.query_data[:5]):
-            try:
-                search_start = time.time()
-                
-                # Combined search: similarity + metadata filter
-                results = self.client.search_vectors(
-                    collection_id=self.collection_name,
-                    query_vectors=[query["vector"]],
-                    top_k=20,
-                    include_metadata=True,
-                    include_vectors=False
-                )
-                
-                search_time = time.time() - search_start
-                
-                # Simulate additional metadata filtering
-                # In real system, this would be done server-side
-                category_filter = query["category"]
-                filtered_results = []
-                for result in results:
-                    # Simulate metadata matching
-                    if len(filtered_results) < 10:  # Simulate finding matches
-                        filtered_results.append(result)
-                
-                combined_metrics = {
-                    "query_id": query["query_id"],
-                    "category_filter": category_filter,
-                    "search_time_ms": search_time * 1000,
-                    "total_results": len(results),
-                    "filtered_results": len(filtered_results),
-                    "top_score": results[0].score if results else 0
-                }
-                
-                combined_results.append(combined_metrics)
-                
-                print(f"Query {i+1} + {category_filter:>10s} filter: "
-                      f"{len(results):2d} → {len(filtered_results):2d} results in {search_time*1000:.1f}ms | "
-                      f"Top score: {results[0].score:.4f}")
-                
-            except Exception as e:
-                print(f"❌ Combined search {i+1} failed: {e}")
-        
-        self.metrics["search"]["combined_searches"] = combined_results
-        return True
-    
-    def generate_performance_report(self) -> Dict:
-        """Generate comprehensive performance report"""
-        print(f"\n📊 COMPREHENSIVE PERFORMANCE REPORT")
-        print("=" * 40)
-        
-        report = {
-            "benchmark_info": {
-                "corpus_size": len(self.corpus_data),
-                "vector_dimension": self.dimension,
-                "batch_size": self.batch_size,
-                "storage_engine": "VIPER",
-                "indexing": "AXIS",
-                "timestamp": time.time()
-            },
-            "insertion_performance": {},
-            "search_performance": {},
-            "overall_metrics": {}
-        }
-        
-        # Insertion performance
-        insertion = self.metrics["insertion"]
-        successful_batches = [b for b in insertion["batches"] if b["success"]]
-        
-        if successful_batches:
-            batch_times = [b["time_seconds"] for b in successful_batches]
-            batch_rates = [b["rate_vectors_per_sec"] for b in successful_batches]
-            
-            report["insertion_performance"] = {
-                "total_vectors": insertion["total_vectors"],
-                "total_time_seconds": insertion["total_time"],
-                "overall_rate_vectors_per_sec": insertion["total_vectors"] / insertion["total_time"],
-                "successful_batches": len(successful_batches),
-                "failed_batches": insertion["errors"],
-                "batch_time_stats": {
-                    "mean": statistics.mean(batch_times),
-                    "min": min(batch_times),
-                    "max": max(batch_times),
-                    "stdev": statistics.stdev(batch_times) if len(batch_times) > 1 else 0
-                },
-                "batch_rate_stats": {
-                    "mean": statistics.mean(batch_rates),
-                    "min": min(batch_rates),
-                    "max": max(batch_rates),
-                    "stdev": statistics.stdev(batch_rates) if len(batch_rates) > 1 else 0
-                }
-            }
-        
-        # Search performance
-        search = self.metrics["search"]
-        
-        # Similarity search stats
-        sim_searches = [s for s in search["similarity_searches"] if s.get("success", True)]
-        if sim_searches:
-            sim_times = [s["search_time_ms"] for s in sim_searches]
-            found_count = sum(1 for s in sim_searches if s["expected_found"])
-            
-            report["search_performance"]["similarity_search"] = {
-                "total_searches": len(sim_searches),
-                "expected_matches_found": found_count,
-                "accuracy_rate": found_count / len(sim_searches),
-                "search_time_stats_ms": {
-                    "mean": statistics.mean(sim_times),
-                    "min": min(sim_times),
-                    "max": max(sim_times),
-                    "stdev": statistics.stdev(sim_times) if len(sim_times) > 1 else 0
-                }
-            }
-        
-        # Print summary
-        print(f"🎯 INSERTION PERFORMANCE:")
-        ins_perf = report["insertion_performance"]
-        print(f"   • Total vectors: {ins_perf['total_vectors']:,}")
-        print(f"   • Total time: {ins_perf['total_time_seconds']:.2f}s")
-        print(f"   • Overall rate: {ins_perf['overall_rate_vectors_per_sec']:.0f} vectors/sec")
-        print(f"   • Batch success rate: {ins_perf['successful_batches']}/{ins_perf['successful_batches']+ins_perf['failed_batches']}")
-        
-        if "similarity_search" in report["search_performance"]:
-            print(f"\n🎯 SEARCH PERFORMANCE:")
-            search_perf = report["search_performance"]["similarity_search"]
-            print(f"   • Search accuracy: {search_perf['accuracy_rate']*100:.1f}%")
-            print(f"   • Avg search time: {search_perf['search_time_stats_ms']['mean']:.1f}ms")
-            print(f"   • Search time range: {search_perf['search_time_stats_ms']['min']:.1f}-{search_perf['search_time_stats_ms']['max']:.1f}ms")
-        
-        # Save report
-        with open("bert_benchmark_report.json", "w") as f:
-            json.dump(report, f, indent=2)
-        
-        print(f"\n💾 Full report saved to: bert_benchmark_report.json")
-        
-        return report
-    
-    async def cleanup(self):
-        """Clean up resources"""
-        if self.client:
-            await self.client.close()
-        
-        if self.server_process:
-            self.server_process.terminate()
-            try:
-                await asyncio.wait_for(self.wait_for_process(), timeout=10.0)
-            except asyncio.TimeoutError:
-                self.server_process.kill()
-                await self.wait_for_process()
-    
-    async def wait_for_process(self):
-        """Wait for server process to terminate"""
-        while self.server_process.poll() is None:
-            await asyncio.sleep(0.1)
-
-async def main():
-    print("🧠 BERT + VIPER Comprehensive Benchmark")
-    print("🎯 10K vectors, gRPC batching, AXIS indexing, metadata filtering")
-    print("=" * 80)
-    
-    benchmark = BERTViperBenchmark()
-    
-    try:
-        # Load data
-        if not await benchmark.load_corpus_data():
-            return 1
-        
-        # Start server
-        if not await benchmark.start_server():
-            return 1
-        
-        # Setup client
-        if not await benchmark.setup_client():
-            return 1
-        
-        # Create collection
-        if not await benchmark.create_viper_collection():
-            return 1
-        
-        # Insert vectors
-        if not await benchmark.insert_vectors_in_batches():
-            print("⚠️ Insertion had issues, but continuing...")
-        
-        # Wait for indexing
-        if not await benchmark.wait_for_indexing():
-            print("⚠️ Indexing verification failed, but continuing...")
-        
-        # Run search tests
-        await benchmark.test_similarity_search()
-        await benchmark.test_id_search()
-        await benchmark.test_metadata_filtering()
-        await benchmark.test_combined_search()
-        
-        # Generate report
-        report = benchmark.generate_performance_report()
-        
-        print(f"\n🎉 BENCHMARK COMPLETE!")
-        print(f"📊 Successfully tested BERT embeddings with VIPER engine")
-        print(f"🚀 Full layered search system verified")
-        
-        return 0
-        
-    except Exception as e:
-        print(f"💥 Benchmark failed: {e}")
-        traceback.print_exc()
-        return 1
-    finally:
-        await benchmark.cleanup()
-
-if __name__ == "__main__":
-    exit_code = asyncio.run(main())
-    sys.exit(exit_code)
