--- search_optimization_test_results.json
@@ -1,4 +1,0 @@
-{
-  "viper_tests": [],
-  "lsm_tests": []
-}
