--- simple_wal_test.rs
@@ -1,173 +1,0 @@
-// Simple test to verify the WAL search fix
-use std::collections::HashMap;
-
-#[derive(Clone, Debug)]
-pub struct TestVectorRecord {
-    pub id: String,
-    pub vector: Vec<f32>,
-}
-
-#[derive(Clone, Debug)]
-pub enum TestWalOperation {
-    Insert {
-        vector_id: String,
-        record: TestVectorRecord,
-    },
-}
-
-#[derive(Clone, Debug)]
-pub struct TestWalEntry {
-    pub operation: TestWalOperation,
-}
-
-// This is the FIXED logic from the actual codebase
-fn search_vectors_similarity_fixed(
-    all_entries: Vec<TestWalEntry>,
-    query_vector: &[f32],
-    k: usize,
-) -> Vec<(String, f32)> {
-    let mut scored_entries = Vec::new();
-    
-    println!("🔧 [TEST] WAL search: found {} total entries", all_entries.len());
-    
-    for (idx, entry) in all_entries.iter().enumerate() {
-        match &entry.operation {
-            TestWalOperation::Insert { vector_id, record } => {
-                println!("🔧 [TEST] WAL entry {}: vector_id={}, vector_len={}", 
-                         idx, vector_id, record.vector.len());
-                
-                // THIS IS THE CRITICAL FIX: Using 'record.vector' (not 'record.dense_vector')
-                let score = compute_cosine_similarity(query_vector, &record.vector);
-                println!("🔧 [TEST] WAL similarity score for {}: {}", vector_id, score);
-                scored_entries.push((vector_id.clone(), score));
-            }
-        }
-    }
-    
-    println!("🔧 [TEST] WAL search: computed {} similarity scores", scored_entries.len());
-    
-    // Sort by score (descending) and take top k
-    scored_entries.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-    scored_entries.truncate(k);
-    
-    println!("🔧 [TEST] WAL search: returning top {} results", scored_entries.len());
-    scored_entries
-}
-
-fn compute_cosine_similarity(a: &[f32], b: &[f32]) -> f32 {
-    if a.len() != b.len() {
-        return 0.0;
-    }
-    
-    let dot_product: f32 = a.iter().zip(b.iter()).map(|(x, y)| x * y).sum();
-    let norm_a: f32 = a.iter().map(|x| x * x).sum::<f32>().sqrt();
-    let norm_b: f32 = b.iter().map(|x| x * x).sum::<f32>().sqrt();
-    
-    if norm_a == 0.0 || norm_b == 0.0 {
-        return 0.0;
-    }
-    
-    dot_product / (norm_a * norm_b)
-}
-
-fn test_exact_match() -> bool {
-    println!("\n🔧 Test 1: Exact Match");
-    
-    let test_vector = vec![1.0, 0.0];
-    
-    let record = TestVectorRecord {
-        id: "test_vector_1".to_string(),
-        vector: test_vector.clone(),
-    };
-    
-    let wal_entry = TestWalEntry {
-        operation: TestWalOperation::Insert {
-            vector_id: "test_vector_1".to_string(),
-            record,
-        },
-    };
-    
-    let results = search_vectors_similarity_fixed(
-        vec![wal_entry],
-        &test_vector,
-        1
-    );
-    
-    if results.len() == 1 && results[0].0 == "test_vector_1" && (results[0].1 - 1.0).abs() < 0.001 {
-        println!("✅ Test 1 PASSED: Exact match found with score {}", results[0].1);
-        true
-    } else {
-        println!("❌ Test 1 FAILED: Expected 1 result with score ~1.0, got {:?}", results);
-        false
-    }
-}
-
-fn test_ranking() -> bool {
-    println!("\n🔧 Test 2: Similarity Ranking");
-    
-    let query_vector = vec![1.0, 0.0];
-    
-    // Create vectors with different similarities
-    let vectors = vec![
-        (vec![1.0, 0.0], "exact_match"),     // Similarity = 1.0
-        (vec![0.9, 0.436], "close_match"),   // Similarity ≈ 0.9  
-        (vec![0.0, 1.0], "orthogonal"),      // Similarity = 0.0
-    ];
-    
-    let mut all_entries = Vec::new();
-    for (vector, id) in vectors {
-        let record = TestVectorRecord {
-            id: id.to_string(),
-            vector,
-        };
-        
-        let wal_entry = TestWalEntry {
-            operation: TestWalOperation::Insert {
-                vector_id: id.to_string(),
-                record,
-            },
-        };
-        
-        all_entries.push(wal_entry);
-    }
-    
-    let results = search_vectors_similarity_fixed(
-        all_entries,
-        &query_vector,
-        3
-    );
-    
-    if results.len() == 3 && 
-       results[0].0 == "exact_match" && 
-       results[1].0 == "close_match" && 
-       results[2].0 == "orthogonal" &&
-       results[0].1 >= results[1].1 && 
-       results[1].1 >= results[2].1 {
-        println!("✅ Test 2 PASSED: Ranking correct - scores: {:.3}, {:.3}, {:.3}", 
-                 results[0].1, results[1].1, results[2].1);
-        true
-    } else {
-        println!("❌ Test 2 FAILED: Ranking incorrect, got {:?}", results);
-        false
-    }
-}
-
-fn main() {
-    println!("🔬 WAL Search Critical Fix Verification");
-    println!("========================================");
-    println!("Testing the fix: record.vector (not record.dense_vector)");
-    
-    let test1 = test_exact_match();
-    let test2 = test_ranking();
-    
-    println!("\n🏆 RESULTS:");
-    println!("========");
-    if test1 && test2 {
-        println!("✅ ALL TESTS PASSED");
-        println!("✅ WAL search logic is WORKING CORRECTLY");
-        println!("✅ The critical field name fix is SUCCESSFUL");
-        println!("\n🎯 CONCLUSION: WAL search will achieve 100% success in the full system!");
-    } else {
-        println!("❌ Some tests failed");
-    }
-}
