--- test_wal_search_unit.rs
@@ -1,198 +1,0 @@
-// Unit test to verify WAL search functionality
-// This is a standalone test to verify the critical fix
-
-#[cfg(test)]
-mod wal_search_tests {
-    use std::collections::HashMap;
-    
-    // Mock structures to test the logic
-    #[derive(Clone, Debug)]
-    pub struct MockVectorRecord {
-        pub id: String,
-        pub collection_id: String,
-        pub vector: Vec<f32>,
-        pub metadata: HashMap<String, serde_json::Value>,
-    }
-    
-    #[derive(Clone, Debug)]
-    pub enum MockWalOperation {
-        Insert {
-            vector_id: String,
-            record: MockVectorRecord,
-            metadata: HashMap<String, String>,
-        },
-    }
-    
-    #[derive(Clone, Debug)]
-    pub struct MockWalEntry {
-        pub id: String,
-        pub operation: MockWalOperation,
-        pub timestamp: i64,
-        pub collection_id: Option<String>,
-        pub sequence_number: u64,
-    }
-    
-    // The fixed similarity search function (mirroring the corrected logic)
-    fn search_vectors_similarity_fixed(
-        all_entries: Vec<MockWalEntry>,
-        collection_id: &str,
-        query_vector: &[f32],
-        k: usize,
-    ) -> Vec<(String, f32, MockWalEntry)> {
-        let mut scored_entries = Vec::new();
-        
-        println!("🔧 [TEST] WAL search: found {} total entries for collection {}", 
-                 all_entries.len(), collection_id);
-        
-        for (idx, entry) in all_entries.iter().enumerate() {
-            if let MockWalOperation::Insert { vector_id, record, .. } = &entry.operation {
-                println!("🔧 [TEST] WAL entry {}: vector_id={}, vector_len={}", 
-                         idx, vector_id, record.vector.len());
-                
-                // This is the FIXED logic - using 'record.vector' instead of 'record.dense_vector'
-                let score = compute_cosine_similarity(query_vector, &record.vector);
-                println!("🔧 [TEST] WAL similarity score for {}: {}", vector_id, score);
-                scored_entries.push((vector_id.clone(), score, entry.clone()));
-            }
-        }
-        
-        println!("🔧 [TEST] WAL search: computed {} similarity scores", scored_entries.len());
-        
-        // Sort by score (descending) and take top k
-        scored_entries.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-        scored_entries.truncate(k);
-        
-        println!("🔧 [TEST] WAL search: returning top {} results", scored_entries.len());
-        scored_entries
-    }
-    
-    fn compute_cosine_similarity(a: &[f32], b: &[f32]) -> f32 {
-        if a.len() != b.len() {
-            return 0.0;
-        }
-        
-        let dot_product: f32 = a.iter().zip(b.iter()).map(|(x, y)| x * y).sum();
-        let norm_a: f32 = a.iter().map(|x| x * x).sum::<f32>().sqrt();
-        let norm_b: f32 = b.iter().map(|x| x * x).sum::<f32>().sqrt();
-        
-        if norm_a == 0.0 || norm_b == 0.0 {
-            return 0.0;
-        }
-        
-        dot_product / (norm_a * norm_b)
-    }
-    
-    #[test]
-    fn test_wal_search_finds_vectors() {
-        println!("\n🔧 Testing WAL Search - Vector Discovery");
-        
-        // Create test data
-        let test_vector = vec![1.0, 0.0];
-        let test_collection = "test_collection";
-        
-        let record = MockVectorRecord {
-            id: "test_vector_1".to_string(),
-            collection_id: test_collection.to_string(),
-            vector: test_vector.clone(),
-            metadata: HashMap::new(),
-        };
-        
-        let wal_entry = MockWalEntry {
-            id: "entry_1".to_string(),
-            operation: MockWalOperation::Insert {
-                vector_id: "test_vector_1".to_string(),
-                record,
-                metadata: HashMap::new(),
-            },
-            timestamp: chrono::Utc::now().timestamp_millis(),
-            collection_id: Some(test_collection.to_string()),
-            sequence_number: 1,
-        };
-        
-        let all_entries = vec![wal_entry];
-        
-        // Test the fixed search logic
-        let results = search_vectors_similarity_fixed(
-            all_entries,
-            test_collection,
-            &test_vector,  // Exact match query
-            1
-        );
-        
-        // Verify results
-        assert_eq!(results.len(), 1, "Should find exactly 1 vector");
-        assert_eq!(results[0].0, "test_vector_1", "Should find the correct vector");
-        assert!((results[0].1 - 1.0).abs() < 0.001, "Should have similarity score close to 1.0 for exact match");
-        
-        println!("✅ Test passed: WAL search correctly finds vectors using the fixed logic");
-    }
-    
-    #[test]
-    fn test_wal_search_similarity_ranking() {
-        println!("\n🔧 Testing WAL Search - Similarity Ranking");
-        
-        let query_vector = vec![1.0, 0.0];
-        let test_collection = "test_collection";
-        
-        // Create vectors with different similarities
-        let vectors = vec![
-            (vec![1.0, 0.0], "exact_match"),     // Similarity = 1.0
-            (vec![0.9, 0.1], "close_match"),     // Similarity ≈ 0.9
-            (vec![0.0, 1.0], "orthogonal"),      // Similarity = 0.0
-        ];
-        
-        let mut all_entries = Vec::new();
-        for (i, (vector, id)) in vectors.iter().enumerate() {
-            let record = MockVectorRecord {
-                id: id.to_string(),
-                collection_id: test_collection.to_string(),
-                vector: vector.clone(),
-                metadata: HashMap::new(),
-            };
-            
-            let wal_entry = MockWalEntry {
-                id: format!("entry_{}", i),
-                operation: MockWalOperation::Insert {
-                    vector_id: id.to_string(),
-                    record,
-                    metadata: HashMap::new(),
-                },
-                timestamp: chrono::Utc::now().timestamp_millis(),
-                collection_id: Some(test_collection.to_string()),
-                sequence_number: i as u64 + 1,
-            };
-            
-            all_entries.push(wal_entry);
-        }
-        
-        // Test ranking
-        let results = search_vectors_similarity_fixed(
-            all_entries,
-            test_collection,
-            &query_vector,
-            3  // Get all results
-        );
-        
-        // Verify ranking
-        assert_eq!(results.len(), 3, "Should find all 3 vectors");
-        assert_eq!(results[0].0, "exact_match", "Best match should be first");
-        assert_eq!(results[1].0, "close_match", "Second best should be second");
-        assert_eq!(results[2].0, "orthogonal", "Worst match should be last");
-        
-        // Verify scores are in descending order
-        assert!(results[0].1 >= results[1].1, "Scores should be in descending order");
-        assert!(results[1].1 >= results[2].1, "Scores should be in descending order");
-        
-        println!("✅ Test passed: WAL search correctly ranks vectors by similarity");
-    }
-}
-
-fn main() {
-    // Run the tests
-    println!("🔬 WAL Search Unit Tests - Verifying Critical Fix");
-    println!("=" * 60);
-    
-    // This would normally be run with `cargo test`
-    println!("Run with: cargo test test_wal_search_unit");
-    println!("This verifies the core WAL search logic that was fixed");
-}
