--- verify_wal_config.py
@@ -1,85 +1,0 @@
-#!/usr/bin/env python3
-
-"""
-Simple verification that WAL configuration is working correctly.
-"""
-
-import os
-from pathlib import Path
-
-def main():
-    print("🔧 WAL Configuration Verification")
-    print("=" * 50)
-    
-    # 1. Check config.toml has new WAL configuration
-    config_path = "/workspace/config.toml"
-    if not os.path.exists(config_path):
-        print("❌ config.toml not found")
-        return 1
-    
-    # Simple text search for WAL config
-    with open(config_path, 'r') as f:
-        config_content = f.read()
-    
-    if '[storage.wal_config]' in config_content:
-        print("✅ WAL configuration section found in config.toml")
-        
-        # Extract key configuration lines
-        lines = config_content.split('\n')
-        wal_section = False
-        for line in lines:
-            line = line.strip()
-            if line == '[storage.wal_config]':
-                wal_section = True
-                continue
-            elif line.startswith('[') and wal_section:
-                break
-            elif wal_section and '=' in line:
-                print(f"   • {line}")
-    else:
-        print("❌ No [storage.wal_config] section found in config.toml")
-        return 1
-    
-    # 2. Check WAL directory exists and has data
-    wal_path = Path("/workspace/data/wal")
-    if not wal_path.exists():
-        print("❌ WAL directory does not exist")
-        return 1
-    
-    # List collection directories
-    collection_dirs = [d for d in wal_path.iterdir() if d.is_dir()]
-    print(f"\n📁 WAL Directory: {wal_path}")
-    print(f"   • Collection directories: {len(collection_dirs)}")
-    
-    total_wal_files = 0
-    total_size = 0
-    
-    for coll_dir in collection_dirs:
-        wal_files = list(coll_dir.glob("*.avro"))
-        if wal_files:
-            total_wal_files += len(wal_files)
-            for wal_file in wal_files:
-                size = wal_file.stat().st_size
-                total_size += size
-                print(f"   • {coll_dir.name}/wal_current.avro: {size:,} bytes")
-    
-    print(f"\n📊 Summary:")
-    print(f"   • Total collections with WAL data: {len(collection_dirs)}")
-    print(f"   • Total WAL files: {total_wal_files}")
-    print(f"   • Total WAL data size: {total_size:,} bytes ({total_size/1024/1024:.2f} MB)")
-    
-    if total_wal_files > 0:
-        print("\n✅ SUCCESS: WAL configuration is working correctly!")
-        print("🎉 Evidence:")
-        print("   • Configuration loaded from config.toml with wal_config section")
-        print("   • URL-based WAL paths (file:///workspace/data/wal)")
-        print("   • Collection-specific WAL directories created")
-        print("   • Avro WAL files contain actual data")
-        print("   • No hardcoded paths in use")
-        return 0
-    else:
-        print("\n⚠️  WARNING: No WAL files found (server may not have been used yet)")
-        return 0
-
-if __name__ == "__main__":
-    exit(main())
